==> 01-clean.Rmd <==
---
title: '01. Cleaning experimental observations'
author: "Al Xin"
date: "August 2020"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = F)
if (!require(pacman)) install.packages("pacman")
library(pacman)
```

## Overview

In this step, typos will be removed and the conditions for the experimental data will be standardized (298 K, water as a solvent, neutral pH, etc.) The cleaned data will be stored in the directory `affinity/clean`.

## Dependencies

Packages needed in this step include `data.table`, `dplyr`, `stringr`, and `tidyr`. Additionally, a personal function from `qsarr` for regular expressions is also used. 

```{r packages, message = F, warning = F}
if(!dir.exists("affinity/clean")) dir.create("affinity/clean")
# devtools::install_github("awqx/qsarr")
library(qsarr)
library(data.table)
library(dplyr)
library(stringr)
library(tidyr)
```

## Rekharsky and Inoue dataset

The data is read from the previous step. The data contains the following columns:
- `host`: α-, β-, or γ-cyclodextrin (written as `"alpha", "beta", "gamma"`) 
- `guest`: the ligand tested in the complex
- `dG`: the change in Gibbs free energy of the complexation, in kJ/mol
- `solvent`: a description of the solvent tested
- `temp`: the temperature of the experiment
- `ref`: the reference from the paper corresponding to the test

### Solvent conditions 

For consistency across observations, the data used will contain only observations conducted at 298 K in water as a solvent. Additionally, observations will be filtered to include observations from pH 6.8 to 7.2.

The original dataset contains 1367 observations. After cleaning for solvent conditions, the dataset contains 802 observations.

```{r ri_solvent}
# Reading the data from the previous step
ri_df <- readRDS("affinity/derived/ri-df.RDS")

# Removing unusual separators and spaces
ri <- ri_df %>%
  lapply(
    str_replace_all, pattern = "(\\â)|\u2212",
    replacement = "-") %>%
  lapply(str_replace_all, "(\\â+)|(Â·)|·|\u2009", " ") %>%
  lapply(str_replace_all, "\\s+", " ") %>%
  lapply(str_replace_all, "\\- ", "-") %>%
  as_tibble() %>%
  mutate(guest = str_replace_all(guest, "\\.", ","))

# Creating a column for pH
pH_regex <-  "[0-9]+\\.*[0-9]*"

# Splitting columns for solvent and solvent description
ri <- separate(
  ri,
  solvent, 
  c("solvent","solvent_desc"),
  sep = "(?=\\s*\\()", 
  extra = "merge", 
  fill = "right"
  )
ri <- ri %>% 
  mutate(
    pH = str_extract(
      solvent_desc, 
      "(pH\\s(\\<\\s)*[0-9]+(\\.[0-9]+)*)"),
    pH_range = str_extract(
      solvent_desc,
      "pH\\s[0-9]+\\.[0-9]+\\-[0-9]+(\\.[0-9]+)*")) %>%
  mutate(pH = ifelse(!is.na(pH_range), NA, pH)) %>%
  mutate(pH = as.numeric(str_extract(pH, pH_regex))) %>%
  separate(., pH_range, c("pH1", "pH2"), sep = "-") %>%
  mutate(
    pH1 = as.numeric(
      str_extract(pH1, pattern = pH_regex)),
    pH2 = as.numeric(
      str_extract(pH2, pattern = pH_regex))) %>%
  mutate(pH = ifelse(!is.na(pH1), (pH1 + pH2)/2, pH)) %>%
  mutate(pH = ifelse(is.na(pH), 7.0, pH)) %>%
  select(-pH1, -pH2) 


# Cleaning for temperature and solvent
ri <- ri %>%
  mutate(temp = as.numeric(temp)) %>%
  filter(
    solvent == "H2O", 
    temp == 298, 
    6.8 <= pH & pH <= 7.2
  )
```

### Data reliability

According to the original Rekharsky and Inoue paper, datapoints with the footnote b, c, g, i, j, or m contain data that may be unreliable. These datapoints will be cleaned from the set. This leaves 774 observations.

```{r ri_ref}
ri <- filter(ri, !str_detect(ref, "b|c|g|i|j|m"))
```

### Removing salts

Hydrochloride salts will also be removed from the dataset. These are guest ligands that contain `HCl` in the name. This leaves 708 observations. 

```{r ri_salt}
ri <- filter(ri, !str_detect(guest, "HCl$"))
```

### Gibbs free energy data

Estimates of Gibbs free energy containing margins of error (e.g., plus or minus a measure) are cleaned to only contain the central estimate. 

Any observations with no observation for Gibbs free energy will be removed. 

Additionally, ligands with multiple reprted Gibbs free energy observations will be summarized by the average of the observations. This leaves 561 observations.

```{r ri_dg}
ri <- ri %>%
  mutate(dG = str_remove(dG, "\\s\\±\\s.+")) %>%
  mutate(dG = as.numeric(dG)) %>%
  group_by(host, guest) %>%
  summarize(dG = mean(dG))

ri <- filter(ri, !is.na(dG))
```

### Cyclodextrin hosts

This study will only examine α-, β-, or γ-cyclodextrin. Other varieties of cyclodextrin will be removed from the dataset. This leaves 470 observations.

```{r ri_cd}
ri <- filter(
  ri, 
  str_detect(host, "^(alpha|beta|gamma)"))
```

```{r ri_cd_cat, echo = F}
cat(
  "a-CD observations:", 
  sum(str_count(ri$host, "alpha")), 
  "\nb-CD:", 
  sum(str_count(ri$host, "beta")), 
  "\nc-CD:", 
  sum(str_count(ri$host, "gamma")))
```

### Name correction 

There are ligands that contain a description of whether the ligand is an anion, monoanion, or dianion. These descriptions must be removed before passing the names into the search engine in `02-sdf.Rmd`.

```{r ri_ion}
ri <- mutate(
  ri, 
  guest = str_remove(
    guest, 
    "\\(anion\\)|\\(monoanion\\)|\\(dianion\\)"))
```

Additionally, some ligands have typos that can only be corrected through manual replacement. The file containing these replacements is `ligand_name.csv`. The replacements must be converted into regular expressions for `str_replace`. 

```{r ri_replace}
ri_replace <- read.csv("ligand_name.csv", header = T) %>%
  mutate(
    pattern = as.character(pattern), 
    replacement = as.character(replacement)) %>%
  mutate(
    pattern = paste0(make_regex(pattern), "$"), 
    replacement = make_regex(replacement)
  )
```

The incorrect names can be iteratively replaced. A `for` loop is used here instead of `mapply` because the argument for replacement needs to be the result of the last function call. This results in 468 observations

```{r ri_replace_df}
for (i in 1:nrow(ri_replace)) {
  ri$guest <- str_replace(
    ri$guest, 
    pattern = ri_replace[i, 1], 
    replacement = ri_replace[i, 2]
  )
}

# Certain duplicated guests have to be collapsed
ri <- ri %>%
  group_by(host, guest) %>%
  summarize(dG = mean(dG))
```

### Saving data

The relevant columns for later steps are `host`, `guest`, and `dG`. These will be saved in `affinity/clean`. In total, 1367 observations were cleaned to 468 observations.

```{r ri_save}
ri %>%
  select(host, guest, dG) %>%
  data.frame() %>%
  saveRDS("affinity/clean/ri.RDS")
```

## Suzuki dataset

The dataset from Suzuki (2001) requires less cleaning as the conditions for the experiment are already standardized. The dataset contains three columns:

- `host`: α-, β-, or γ-cyclodextrin (written as `"alpha", "beta", "gamma"`) 
- `guest`: the ligand tested in the complex
- `dG`: the change in Gibbs free energy of the complexation, in kJ/mol

Observations without reported change in Gibbs free energy are removed. Out of 436 original observations, 320 remain after cleaning. The cleaned data will be saved in `affinity/clean`.

```{r suzuki_clean}
# Replace non-standard hyphen and space
readRDS("affinity/derived/suzuki-df.RDS") %>%
  mutate(dG = str_replace(dG, "\u2212", "-")) %>%
  mutate(dG = as.numeric(dG)) %>%
  mutate(guest = str_replace(guest, "\u2009", " ")) %>%
  filter(!is.na(dG)) %>%
  saveRDS(suzuki, "affinity/clean/suzuki.RDS")
```

## Singh et al dataset

The Singh et al (2015) dataset, like the Suzuki information, is largely cleaned and only needs to be filtered for reported Gibbs free energy change. A slight error occurred in the previous step, where the host, which should be all "gamma" was coerced to "alpha" through collision of the factor levels. This can be corrected manually. Additionally, the names of the guest ligands need to be converted to lowercase. Out of 114 observations, 112 are retained. The results will be saved in `affinity/clean`.

```{r singh_clean}
readRDS("affinity/derived/singh-df.RDS") %>% 
  filter(!is.na(dG)) %>% 
  mutate(guest = tolower(guest)) %>%
  mutate(host = "gamma") %>%
  saveRDS("affinity/clean/singh.RDS")
```

## Combined dataset

The data from all three sources are also compiled into a single data frame. For repeated observations (same host and guest), the average change in Gibbs free energy is reported. The compiled data will be abreviated as `riss`.

There are 909 observations between the three data sets. Taking the average Gibbs free energy summarizes this to 766 observations.

```{r riss}
ri <- readRDS("affinity/clean/ri.RDS")
suzuki <- readRDS("affinity/clean/suzuki.RDS")
singh <- readRDS("affinity/clean/singh.RDS")
riss <- rbind(ri, suzuki, singh) %>%
  group_by(guest, host) %>%
  summarize(dG = mean(dG))
saveRDS(riss, "affinity/clean/riss.RDS")
```
