==> 05-join.Rmd <==
---
title: "05. Joining affinity data to descriptors"
author: "Al Xin"
date: "12/31/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = F)
```

## Overview

The binding affinity in the `affinity` data will be joined to the chemical descriptors. The joined dataset will consist of the chemical descriptors with corresponding Gibbs free energy change. This data will be saved in `join/`. The directory structure will be the same as `desc/`.

```{r dir_create}
src_list <- c("cdk", "mordred", "ochem", "padel")
dir_list <- paste0("join/", src_list)
if (!dir.exists("join")) {
  dir_result <- lapply(c("join", dir_list), dir.create)
}
```

## Dependencies

```{r}
library(dplyr)
library(qsarr)
library(stringr)
```


## Attaching guest names

The observed Gibbs free energy values can be read from `affinity/clean`. Calling `dplyr::inner_join` with the chemical descriptor data completes the data we need for training. However, there are some data sets that are missing the guest names and these need to be replaced.

### Mordred

For descriptors generated by Mordred, the name has been removed due to incompatibility with commas in the written CSV. These names have to be replaced before joining the data frames.

```{r mordred}
cd_list <- c("alpha", "beta", "gamma")
df_list <- c("ri", "suzuki", "singh", "riss")
dir_list <- paste0("sdf/", df_list)
dir_list <- paste0(
  dir_list, "/", 
  unlist(lapply(cd_list, rep, 4)))
desc_name <- dir_list %>% 
  str_remove("sdf/") %>%
  str_replace("/", "_")

# ri_alpha <- readRDS("trn/mordred/ri_alpha.RDS") %>%
#   mutate(name = as.integer(name))
# guest_list <- list.files("sdf/ri/alpha") %>%
#   str_remove("\\.SDF")
# guest_name <- guest_list[ri_alpha$name]

guest_list <- lapply(
  dir_list, 
  function(x) {
    list.files(x) %>%
      str_remove("(?i)\\.SDF")
  }
)

names(guest_list) <- desc_name
guest_list <- guest_list[sapply(guest_list, length) > 0]

mordred_list <- read_desc_list(list.files("desc/mordred", full.names = T))

mordred_name_temp <- names(mordred_list)
mordred_list <- lapply(
  names(mordred_list), 
  function(x) {
    df <- mordred_list[[x]]
    guest <- guest_list[[x]]
    guest <- guest[df$name]
    df <- df %>% 
      mutate(name = guest) %>%
      rename(guest = name)
  }
)
names(mordred_list) <- mordred_name_temp
```

### OCHEM

Like Mordred, the descriptors from OCHEM also do not have the names of the guests. The names must be added using `dplyr::mutate`.

For indexing the guest names, `mordred_list` is recycled as the ordering of guests is the same in both cases. In OCHEM, the indexing trick to populate the names for the Mordred descriptors cannot be used because there is no data on names.

```{r ochem, message = F}
ochem_list <- read_desc_list(list.files("desc/ochem", full.names = T)) 

ochem_list <- retain_name(
  ochem_list, 
  function(x) {
    lapply(
      x, 
      as.numeric
    ) %>%
      data.frame()
  }
)

ochem_name_temp <- names(ochem_list)
ochem_list <- lapply(
  names(ochem_list), 
  function(x) {
    guest <- mordred_list[[x]]$guest
    df <- ochem_list[[x]]
    mutate(df, guest = guest)
  }
)
names(ochem_list) <- ochem_name_temp
```

## Loading descriptor data

For descriptors generated by PaDEL-Descriptor or through CDK using `rcdk`, the descriptors can be loaded without additional attaching of guest names. 

```{r}
cdk_list <- read_desc_list(
  list.files("desc/cdk", full.names = T), 
  quiet = T
)

padel_list <- read_desc_list(
  list.files("desc/padel", full.names = T), 
  quiet = T
) %>%
  retain_name(
    x = ., 
    f = function(x) rename(x, guest = Name)
    )
```

## Attaching affinity data

The change in Gibbs free energy is read from the directory of cleaned affinity data and labeled with the name of the data source. 

This allows the function`attach_dg` to easily find the correct affinity when given the name of the descriptor files.

```{r}
dg_list <- read_desc_list(
  list.files("affinity/clean", full.names = T), 
  quiet = T
)
```

Using a series of `lapply` functions, the descriptor lists can be attached to the correct affinity values. 

```{r attach_dg}
attach_dg <- function(df_list, d_name) {
  data_source <- str_extract(d_name, "^[[:alpha:]]+") 
  host_type <- str_extract(d_name, "[[:alpha:]]+$")

  df <- df_list[[d_name]]
  dg <- dg_list[[data_source]] %>%
    filter(host == host_type)
  inner_join(df, dg, by = "guest") %>%
    select(-host)
}

attach_dg_list <- function(df_list) {
  df_name <- names(df_list)
  result <- lapply(
    names(df_list), 
    attach_dg, 
    df_list = df_list
  )
  names(result) <- df_name
  result
}

desc_list <- list(
  cdk_list, 
  mordred_list, 
  ochem_list, 
  padel_list
)

join_list <- lapply(
  desc_list, 
  attach_dg_list
)
```

The resulting lists can then be saved to the appropriate directory by calling a `Map` function. 

```{r join_save}
join_save <- Map(
  function(dir_name, df_list) {
    lapply(
      names(df_list), 
      function(x) {
        df <- df_list[[x]]
        saveRDS(df, paste0(dir_name, "/", x, ".RDS"))
      }
    )
  }, 
  paste0("join/", src_list), 
  join_list
)
```
