==> 11-svm_linear.Rmd <==
---
title: "11. SVM"
author: "Al Xin"
date: "1/5/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, eval = F)
```

**THIS IS A WORK IN PROGRESS. THE `qsarr` FUNCTION FOR SVMs ARE STILL NOT FULLY OPERATIONAL.**

## Overview

Models built on all predictor variables will be saved in `model/svm/all_var/`. 
Models trained on variables selected from RFE in 09 will be saved in `model/svm/rfe_var`. 

```{r directories}
desc_src <- c("cdk", "mordred", "ochem", "padel")

if (!dir.exists("model")) dir.create("model")
if (!dir.exists("model/svm_linear")) {
  dir.create("model/svm_linear")
  dir.create("model/svm_linear/all_var")
  dir.create("model/svm_linear/rfe_var")
  dir_create_result <- sapply(paste0("model/svm_linear/all_var/", desc_src), dir.create)
  dir_create_result <- sapply(paste0("model/svm_linear/rfe_var/", desc_src), dir.create)
}
```

### General outline for model building

## Dependencies

```{r package}
library(caret)
library(data.table)
library(dplyr)
# # library(stringr)
# library(tibble)
# library(tidyr)
library(qsarr)
library(stringr)
library(tidyverse)
# library(randomForest)
library(parallel)
library(doParallel)

theme_set(theme_bw())
```

## All variables

For creating models with all the variables, the descriptors are read as separate 
lists. This makes it clearer that different possible tuning parameters are used
based on the number of descriptors available. 

```{r cdk_all}
# Read preprocessed descriptors
# Remove guest names for training
cdk_list <- read_desc_list(
  list.files("preprocess/cdk", full.names = T), 
  quiet = T
)

padel_list <- read_desc_list(
  list.files("preprocess/padel", full.names = T), 
  quiet = T
)

mordred_list <- read_desc_list(
  list.files("preprocess/mordred", full.names = T), 
  quiet = T
)

ochem_list <- read_desc_list(
  list.files("preprocess/ochem", full.names = T), 
  quiet = T
) 

data_name <- names(cdk_list)

cdk_all <- lapply(
  cdk_list, 
  tune, 
  method = "svm_linear", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(4, 16, 32, 64), 
  nodesize = c(2, 4, 8)
)

cdk_all_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(cdk_all[[x]], paste0("model/svm_linear/all_var/cdk/", x, ".RDS"))
  }
)

padel_all <- parLapply(
  cl = cl,
  padel_list, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(4, 20, 40, 100), 
  nodesize = c(2, 4, 8)
)

padel_all_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(padel_all[[x]], paste0("model/svm_linear/all_var/padel/", x, ".RDS"))
  }
)

mordred_all <- parLapply(
  cl = cl,
  mordred_list, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(4, 20, 40, 100), 
  nodesize = c(2, 4, 8)
)

mordred_all_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(mordred_all[[x]], paste0("model/svm_linear/all_var/mordred/", x, ".RDS"))
  }
)

ochem_all <- parLapply(
  cl = cl,
  ochem_list, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(4, 20, 60, 150), 
  nodesize = c(2, 4, 8)
)

ochem_all_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(ochem_all[[x]], paste0("model/svm_linear/all_var/ochem/", x, ".RDS"))
  }
)

stopCluster(cl = cl)
```
```{r rfe}
# Read preprocessed descriptors
# Remove guest names for training
cdk_list_rfe <- read_desc_list(
  list.files("rfe/df/cdk", full.names = T), 
  quiet = T
)

padel_list_rfe <- read_desc_list(
  list.files("rfe/df/padel", full.names = T), 
  quiet = T
)

mordred_list_rfe <- read_desc_list(
  list.files("rfe/df/mordred", full.names = T), 
  quiet = T
)

ochem_list_rfe <- read_desc_list(
  list.files("rfe/df/ochem", full.names = T), 
  quiet = T
) 

data_name <- names(cdk_list_rfe)

set.seed(20210114)
cdk_rfe <- lapply(
  cdk_list_rfe, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500,
  replace = c(T, F),
  mtry = c(2, 4, 8, 12),
  nodesize = c(2, 4, 8)
)

cdk_rfe_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(cdk_rfe[[x]], paste0("model/svm_linear/rfe_var/cdk/", x, ".RDS"))
  }
)


padel_rfe <- lapply(
  padel_list_rfe, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(2, 4, 8, 12), 
  nodesize = c(2, 4, 8)
)

padel_rfe_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(padel_rfe[[x]], paste0("model/svm_linear/rfe_var/padel/", x, ".RDS"))
  }
)


mordred_rfe <- lapply(
  mordred_list_rfe, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(2, 4, 8, 12), 
  nodesize = c(2, 4, 8)
)

mordred_rfe_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(mordred_rfe[[x]], paste0("model/svm_linear/rfe_var/mordred/", x, ".RDS"))
  }
)

ochem_rfe <- lapply(
  ochem_list_rfe, 
  tune, 
  method = "rf", 
  resp = "dG",
  ignore_col = "guest",
  nfold = 10,
  nrep = 10,
  ntree = 500, 
  replace = c(T, F),
  mtry = c(2, 4, 8, 12), 
  nodesize = c(2, 4, 8)
)

ochem_rfe_save <- lapply(
  data_name, 
  function(x) {
    saveRDS(ochem_rfe[[x]], paste0("model/svm_linear/rfe_var/ochem/", x, ".RDS"))
  }
)
```

## External validation

```{r}
desc_name <- c("cdk", "mordred", "ochem", "padel")
data_name <- str_remove(list.files("desc/cdk"), "\\.RDS$")
```

First, the data generated from Step 07 is read into a list.

```{r}
ev_list <- lapply(
  paste0("extval_pp/", desc_name), 
  function(x) {
    lapply(
      list.files(x, full.names = T), 
      readRDS
    ) %>%
      setNames(nm = data_name)
  }
) %>%
  setNames(nm = desc_name)
```

### Models using all variables

There are two model-building methods. First, random forests were generated based on all the available variables. 

```{r}
rf_allvar <- lapply(
  paste0("model/svm_linear/all_var/", desc_name), 
  function(x) {
    lapply(
      list.files(x, full.names = T), 
      readRDS
    ) %>%
      setNames(nm = data_name)
  }
) %>%
  setNames(nm = desc_name)
```

The results on external validation data can then be predicted. Both the results on all molecules as well as only molecules within the applicability domain are calculated. 

```{r}
ev_allvar <- lapply(
  desc_name, 
  function(x) {
    tune_sublist <- rf_allvar[[x]]
    ev_sublist <- ev_list[[x]]
    lapply(
      data_name, 
      function(y) {
        pred <- predict(tune_sublist[[y]], ev_sublist[[y]])
        data.frame(
          guest = ev_sublist[[y]]$guest, 
          in_ad = ev_sublist[[y]]$in_ad,
          obs = ev_sublist[[y]]$dG, 
          pred = pred
        )
      }
    ) %>%
      setNames(data_name)
  }
) %>% 
  setNames(desc_name)

# All results, regardless of applicability domain
ev_allvar_result <- lapply(
  ev_allvar, 
  function(x) {
    lapply(
      x, 
      defaultSummary
    ) %>%
      setNames(data_name)
  }
) %>%
  setNames(desc_name)

# Accuracy only on applicability domain
ev_allvar_result_ad <- lapply(
  ev_allvar, 
  function(x) {
    lapply(
      x, 
      function(y) {
        y <- y[y$in_ad, ]
        defaultSummary(y)
      }
    ) %>%
      setNames(data_name)
  }
) %>% 
  setNames(desc_name)
```

The same procedure for prediction is carried out with models only built on variables selected with RFE. 

```{r}
rf_rfe <- lapply(
  paste0("model/svm_linear/rfe_var/", desc_name), 
  function(x) {
    lapply(
      list.files(x, full.names = T), 
      readRDS
    ) %>%
      setNames(nm = data_name)
  }
) %>%
  setNames(nm = desc_name)

ev_rfe <- lapply(
  desc_name, 
  function(x) {
    tune_sublist <- rf_rfe[[x]]
    ev_sublist <- ev_list[[x]]
    lapply(
      data_name, 
      function(y) {
        pred <- predict(tune_sublist[[y]], ev_sublist[[y]])
        data.frame(
          guest = ev_sublist[[y]]$guest, 
          in_ad = ev_sublist[[y]]$in_ad,
          obs = ev_sublist[[y]]$dG, 
          pred = pred
        )
      }
    ) %>%
      setNames(data_name)
  }
) %>% 
  setNames(desc_name)

# All results, regardless of applicability domain
ev_rfe_result <- lapply(
  ev_rfe, 
  function(x) {
    lapply(
      x, 
      defaultSummary
    ) %>%
      setNames(data_name)
  }
) %>%
  setNames(desc_name)

# Accuracy only on applicability domain
ev_rfe_result_ad <- lapply(
  ev_rfe, 
  function(x) {
    lapply(
      x, 
      function(y) {
        y <- y[y$in_ad, ]
        defaultSummary(y)
      }
    ) %>%
      setNames(data_name)
  }
) %>% 
  setNames(desc_name)
```

## Results

### Model-building process

The results are first read from the lists generated from tuning and saved in the
directory `result/trn/`.

```{r}
dir.create("result/trn/all_var", recursive = T)
dir.create("result/trn/rfe_var")

rf_allvar_df <- rbindlist(
  lapply(
    desc_name, 
    function(x) {
      desc_list <- rf_allvar[[x]]
      rbindlist(
        lapply(
          data_name, 
          function(y) {
            mutate(desc_list[[y]]$result_summary, data_src = y)
          }
        )
      ) %>%
        mutate(desc_src = x)
    }
  )
) 

rf_rfe_df <- rbindlist(
  lapply(
    desc_name, 
    function(x) {
      desc_list <- rf_rfe[[x]]
      rbindlist(
        lapply(
          data_name, 
          function(y) {
            mutate(desc_list[[y]]$result_summary, data_src = y)
          }
        )
      ) %>%
        mutate(desc_src = x)
    }
  )
) 

saveRDS(rf_allvar_df, "result/trn/all_var/rf.RDS")
saveRDS(rf_rfe_df, "result/trn/rfe_var/rf.RDS")
```
```{r, eval = T, echo = F}
rf_allvar_df <- readRDS("result/trn/all_var/rf.RDS")
rf_rfe_df <- readRDS("result/trn/rfe_var/rf.RDS")

ggplot(rf_allvar_df, aes(x = data_src, y = fold_avg, fill = desc_src)) + 
  geom_col(position = "dodge") + 
  facet_grid(summary_stat ~ ., scales = "free") + 
  labs(
    x = "Data source", 
    y = "Average value over repeated 10-fold CV", 
    title = "Results for random forest models on all variables", 
    fill = "Chemical descriptor"
  )

ggplot(rf_rfe_df, aes(x = data_src, y = fold_avg, fill = desc_src)) + 
  geom_col(position = "dodge") + 
  facet_grid(summary_stat ~ ., scales = "free") + 
  labs(
    x = "Data source", 
    y = "Average value over repeated 10-fold CV", 
    title = "Results for random forest models on RFE-selected variables", 
    fill = "Chemical descriptor"
  )
```


### External validation 

```{r}
ev_allvar_df <- data.table::rbindlist(
  lapply(
    unlist(ev_allvar_result, recursive = F), 
    function(x) {
      data.frame(as.list(x))
    }
  )
) %>%
  data.frame(
    desc_src = unlist(lapply(desc_name, rep, length(data_name))), 
    data_src = rep(data_name, length(desc_name))
  ) %>%
  pivot_longer(RMSE:MAE, names_to = "summary_stat")

ev_allvar_ad_df <- data.table::rbindlist(
  lapply(
    unlist(ev_allvar_result_ad, recursive = F), 
    function(x) {
      data.frame(as.list(x))
    }
  )
) %>%
  data.frame(
    desc_src = unlist(lapply(desc_name, rep, length(data_name))), 
    data_src = rep(data_name, length(desc_name))
  ) %>%
  pivot_longer(RMSE:MAE, names_to = "summary_stat")

ggplot(ev_allvar_df, aes(x = data_src, y = value, fill = desc_src)) + 
  geom_col(position = "dodge") + 
  facet_grid(summary_stat ~ ., scales = "free")

ggplot(ev_allvar_ad_df, aes(x = data_src, y = value, fill = desc_src)) + 
  geom_col(position = "dodge") + 
  facet_grid(summary_stat ~ ., scales = "free")
# rf_tst <- rf_allvar[[1]]
# ev_tst <- ev_list[[1]]
# pred_tst <- lapply(
#   names(rf_tst), 
#   function(x) {
#     # rf_model <- rf_tst[[x]]$model
#     # predict(rf_model, ev_tst[[x]])
#     predict(rf_tst[[x]], ev_tst[[x]])
#   }
# )
```

```{r}
print.tune <- function(tune_obj) {
  cat(
    "Model type: ", class(tune_obj$model), "\n\n",
    "Tuned parameters: ", "\n",
    sapply(
      names(tune_obj$param),
      function(x) {
        paste0("\t", x, ": ", tune_obj$param[[x]], "\n")
      }
    ), "\n",
    "Number of predictors: ", length(tune_obj$pred_name), "\n",
    "Predictors (first 10): ",
    tune_obj$pred_name[1:(min(10, length(tune_obj$pred_name)))],
    "\n", "\n",
    "Model performance: ", "\n"
  )
  print(tune_obj$result_summary)
}

predict.tune <- function(tune_obj, ...) {
  # if (length(class(tune_obj$model)) > 1) {
  #   class(tune_obj$model) <-
  #     class(tune_obj$model)[length(class(tune_obj$model))]
  # }
  # best_model <- tune_obj$model
  # UseMethod("predict", best_model)
  predict(tune_obj$model, ...)
}
```
